requirements.txt
janome
mecab-python3